+++ apps/ddaas/src/app/[locale]/layout.tsx
@@ -1,4 +1,4 @@
-import { appConfig, generatedLocales, localPrefixAsNeeded, defaultLocale } from "@/lib/appConfig";
+import { appConfig, generatedLocales, localePrefixAsNeeded, defaultLocale } from "@/lib/appConfig";
 import { getFumaTranslations } from '@third-ui/fuma/server';
 import { NProgressBar } from '@third-ui/main/nprogress-bar';
 import { RootProvider } from "fumadocs-ui/provider/next";
@@ -27,19 +27,19 @@ export async function generateMetadata({
     description: t('webDescription'),
     keywords: t('keywords'),
     alternates: {
-      canonical: `${appConfig.baseUrl}${getAsNeededLocalizedUrl(locale, '/', localPrefixAsNeeded, defaultLocale)}`,
+      canonical: `${appConfig.baseUrl}${getAsNeededLocalizedUrl(locale, '/', localePrefixAsNeeded, defaultLocale)}`,
       languages: {
-        "en": `${appConfig.baseUrl}${getAsNeededLocalizedUrl('en', '/', localPrefixAsNeeded, defaultLocale)}`,
-        "zh": `${appConfig.baseUrl}${getAsNeededLocalizedUrl('zh', '/', localPrefixAsNeeded, defaultLocale)}`,
-        "ja": `${appConfig.baseUrl}${getAsNeededLocalizedUrl('ja', '/', localPrefixAsNeeded, defaultLocale)}`,
-        "ko": `${appConfig.baseUrl}${getAsNeededLocalizedUrl('ko', '/', localPrefixAsNeeded, defaultLocale)}`,
-        "fr": `${appConfig.baseUrl}${getAsNeededLocalizedUrl('fr', '/', localPrefixAsNeeded, defaultLocale)}`,
-        "de": `${appConfig.baseUrl}${getAsNeededLocalizedUrl('de', '/', localPrefixAsNeeded, defaultLocale)}`,
-        "es": `${appConfig.baseUrl}${getAsNeededLocalizedUrl('es', '/', localPrefixAsNeeded, defaultLocale)}`,
-        "it": `${appConfig.baseUrl}${getAsNeededLocalizedUrl('it', '/', localPrefixAsNeeded, defaultLocale)}`,
-        "pt": `${appConfig.baseUrl}${getAsNeededLocalizedUrl('pt', '/', localPrefixAsNeeded, defaultLocale)}`,
-        "tr": `${appConfig.baseUrl}${getAsNeededLocalizedUrl('tr', '/', localPrefixAsNeeded, defaultLocale)}`,
-        "pl": `${appConfig.baseUrl}${getAsNeededLocalizedUrl('pl', '/', localPrefixAsNeeded, defaultLocale)}`,
+        "en": `${appConfig.baseUrl}${getAsNeededLocalizedUrl('en', '/', localePrefixAsNeeded, defaultLocale)}`,
+        "zh": `${appConfig.baseUrl}${getAsNeededLocalizedUrl('zh', '/', localePrefixAsNeeded, defaultLocale)}`,
+        "ja": `${appConfig.baseUrl}${getAsNeededLocalizedUrl('ja', '/', localePrefixAsNeeded, defaultLocale)}`,
+        "ko": `${appConfig.baseUrl}${getAsNeededLocalizedUrl('ko', '/', localePrefixAsNeeded, defaultLocale)}`,
+        "fr": `${appConfig.baseUrl}${getAsNeededLocalizedUrl('fr', '/', localePrefixAsNeeded, defaultLocale)}`,
+        "de": `${appConfig.baseUrl}${getAsNeededLocalizedUrl('de', '/', localePrefixAsNeeded, defaultLocale)}`,
+        "es": `${appConfig.baseUrl}${getAsNeededLocalizedUrl('es', '/', localePrefixAsNeeded, defaultLocale)}`,
+        "it": `${appConfig.baseUrl}${getAsNeededLocalizedUrl('it', '/', localePrefixAsNeeded, defaultLocale)}`,
+        "pt": `${appConfig.baseUrl}${getAsNeededLocalizedUrl('pt', '/', localePrefixAsNeeded, defaultLocale)}`,
+        "tr": `${appConfig.baseUrl}${getAsNeededLocalizedUrl('tr', '/', localePrefixAsNeeded, defaultLocale)}`,
+        "pl": `${appConfig.baseUrl}${getAsNeededLocalizedUrl('pl', '/', localePrefixAsNeeded, defaultLocale)}`,
       }
     },
     icons: [
@@ -70,7 +70,7 @@ export default async function RootLayout({
       <NextIntlClientProvider messages={messages}>
         <body className={cn(montserrat.className)}>
           <NProgressBar />
-          <ClerkProviderClient locale={locale} localPrefixAsNeeded={localPrefixAsNeeded} defaultLocale={defaultLocale}>
+          <ClerkProviderClient locale={locale} localePrefixAsNeeded={localePrefixAsNeeded} defaultLocale={defaultLocale}>
             <RootProvider
                 i18n={{
                   locale: locale,