+++ apps/ddaas/src/components/credit-popover.tsx
@@ -8,7 +8,7 @@ import { moneyPriceConfig } from '@windrun-huaiin/backend-core/lib';
 import { buildInitUserContextFromEntities } from '@windrun-huaiin/backend-core/context'
 import { getTranslations } from 'next-intl/server';
 import { getAsNeededLocalizedUrl } from '@windrun-huaiin/lib';
-import { localPrefixAsNeeded, defaultLocale } from '@/lib/appConfig';
+import { localePrefixAsNeeded, defaultLocale } from '@/lib/appConfig';
 
 interface CreditPopoverProps {
   locale: string;
@@ -91,7 +91,7 @@ export async function CreditPopover({ locale }: CreditPopoverProps) {
   ];
 
   // 按照项目设置来决定是否带上语言前缀
-  const pricingPageBaseUrl = getAsNeededLocalizedUrl(locale, "/pricing",  localPrefixAsNeeded,  defaultLocale);
+  const pricingPageBaseUrl = getAsNeededLocalizedUrl(locale, "/pricing",  localePrefixAsNeeded,  defaultLocale);
 
   const data: CreditOverviewData = {
     totalBalance,