+++ packages/third-ui/src/main/footer.tsx
@@ -18,11 +18,11 @@ interface FooterData {
 
 interface FooterProps {
   locale: string;
-  localPrefixAsNeeded?: boolean;
+  localePrefixAsNeeded?: boolean;
   defaultLocale?: string;
 }
 
-export async function Footer({ locale, localPrefixAsNeeded = true, defaultLocale = 'en' }: FooterProps) {
+export async function Footer({ locale, localePrefixAsNeeded = true, defaultLocale = 'en' }: FooterProps) {
   const tFooter = await getTranslations({ locale, namespace: 'footer' });
   
   const company = safeT(tFooter, 'company', '');
@@ -43,11 +43,11 @@ export async function Footer({ locale, localPrefixAsNeeded = true, defaultLocale
       <footer>
         <div className="w-full flex flex-col items-center justify-center px-4 py-8 space-y-3">
           <div className="flex flex-wrap items-center justify-center gap-x-2 gap-y-2 text-xs sm:text-sm sm:gap-x-6">
-            <Link href={getAsNeededLocalizedUrl(locale, "/legal/terms", localPrefixAsNeeded, defaultLocale)} className="flex items-center space-x-1 hover:underline">
+            <Link href={getAsNeededLocalizedUrl(locale, "/legal/terms", localePrefixAsNeeded, defaultLocale)} className="flex items-center space-x-1 hover:underline">
               <icons.ReceiptText className="h-3.5 w-3.5"/>
               <span>{data.terms}</span>
             </Link>
-            <Link href={getAsNeededLocalizedUrl(locale, "/legal/privacy", localPrefixAsNeeded, defaultLocale)} className="flex items-center space-x-1 hover:underline">
+            <Link href={getAsNeededLocalizedUrl(locale, "/legal/privacy", localePrefixAsNeeded, defaultLocale)} className="flex items-center space-x-1 hover:underline">
               <icons.ShieldUser className="h-3.5 w-3.5"/>
               <span>{data.privacy}</span>
             </Link>