+++ packages/third-ui/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @windrun-huaiin/third-ui
 
+## 12.1.0
+
+### Minor Changes
+
+- fix(locale): `localPrefixAsNeeded` rename to `localePrefixAsNeeded`
+
+### Patch Changes
+
+- Updated dependencies
+  - @windrun-huaiin/base-ui@12.0.0
+  - @windrun-huaiin/lib@12.1.0
+
 ## 12.0.3
 
 ### Patch Changes